+++ src/main/java/bridge/Application.java
@@ -1,10 +1,12 @@
 package bridge;
 
+import bridge.contoller.GameController;
+
 public class Application {
 
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Controller controller = new Controller();
+        GameController controller = new GameController();
         controller.run();
     }
 }
